requirements.txt
requests
wsaccel
websocket-client
